Properly resolve node_modules directory
Prior to this commit the resolution of node_modules was not correct. When the code would reach into the falsy clause `path.resolve` was simply doing a join on the context and "node_modules". Clearly this is not correct.  This commit introduces the node modules path resolution algorithm found in the node standard lib.  Why not just use the standard lib you may ask? Well javascript doesn't have the concept of private mehtods and thus we have the convention of prefixing "private" methods with an underscore. module._nodeModulePaths() is one of these methods.